+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- added tachyons css library
+- added new listing page for fac/staff by category showing cards
 
 ## [1.5.3] - 2017-12-07
 ### Changed